+++ package.json
@@ -29,6 +29,6 @@
   },
   "devDependencies": {
     "knex-cleaner": "^1.3.0",
-    "nodemon": "^2.0.2"
+    "nodemon": "^2.0.3"
   }
 }
